Using browserify to build
