+++ .debug.json
@@ -1,0 +1,4 @@
+{
+    "DEBUG": true,
+    "RELEASE": false
+}
