+++ .npmignore
@@ -2,3 +2,6 @@ node_modules
 gulpfile.js
 .gitignore
 .travis.yml
+.debug.json
+.release.json
+src
