+++ .release.json
@@ -1,0 +1,4 @@
+{
+    "DEBUG": false,
+    "RELEASE": true
+}
