+++ .travis.yml
@@ -16,4 +16,4 @@ cache:
         - node_modules
 
 script:
-    - npm test
+    - npm run build
