+++ src/shader/extractAttributes.js
