+++ src/shader/extractUniforms.js
