+++ src/shader/generateUniformAccessObject.js
